UWP impl of GuidUtility. Still needs work.
Not currently procuding the same GUIDs as the iOS and Andoid version. This is because the System.Security.Crypography namespace is unavailable on UWP. The Windows.Security.Crytography namespace, however, does contain crypto algorthims. But the API is com,pletely different. This implementation is currently not producing the same results as the iOS and Android versions, and it needs to in order for the Azure data sync implementation to work properly.