Implemented data sync error alerts in iOS and Android
When one user commits changes to the cloud and another user has has not yet pulled the latest version of the data and is editing the same data (but a different version), a conflict will occur. The conflict is returned to the app from the cloud service. I've added some simple UI messaging to alert the user to this condition so that they can resolve it by clearing and refreshing the local data store.

A more elegant solution would be better, but this will suffice for now.